requirements.txt
pandas==0.25.3
requests==2.23.0
discord.py==1.4.1
beautifulsoup4==4.9.1
discord==1.0.1
python-dotenv==0.14.0
